Improve dispatcher version arguments
Fixes a problem where the primary functions couldn't be provided with optional command line arguments (thus preventing the downloading of the 'latest' cache when not supplying an argument). This was incompatible with the import order for the map builder which relied on singletons which sourced data from the supplied version directory variable.